LayoutTests/transforms/3d/preserve-3d-not-applied-to-ancestors-expected.txt

PASS child.getBoundingClientRect().width is 100
PASS successfullyParsed is true

TEST COMPLETE
